+++ src/main/java/com/devalrykemes/forumhub/service/TopicService.java
@@ -8,10 +8,13 @@ import com.devalrykemes.forumhub.domain.topic.Topic;
 import com.devalrykemes.forumhub.domain.topic.TopicRequestDto;
 import com.devalrykemes.forumhub.domain.topic.TopicResponseDto;
 import com.devalrykemes.forumhub.domain.user.User;
+import com.devalrykemes.forumhub.repository.CommentRepository;
 import com.devalrykemes.forumhub.repository.ProfileRepository;
 import com.devalrykemes.forumhub.repository.TopicRepository;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -21,6 +24,7 @@ public class TopicService {
     public final TopicRepository topicRepository;
     public final ProfileService profileService;
     public final CourseService courseService;
+    public final CommentRepository commentRepository;
 
     public TopicResponseDto createTopic(TopicRequestDto topicRequestDto) {
         Topic newTopic = new Topic(topicRequestDto);
@@ -36,11 +40,31 @@ public class TopicService {
 
     public void deleteTopicById(Long id) {
         this.getTopicById(id);
-        topicRepository.deleteById(id);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userEmail = authentication.getName();
+        if(topicRepository.existsByTopicIdAndCreatorEmail(id, userEmail)) {
+            topicRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("This topic has not been created by this user!");
+        }
     }
 
     public Topic topicById(Long id) {
         return topicRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Topic not found: " + id));
+    }
+
+    public TopicResponseDto topicSolvedByComment(Long topicId, Long commentId) {
+        this.getTopicById(topicId);
+        commentRepository.findById(commentId);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String userEmail = authentication.getName();
+        if(topicRepository.existsByTopicIdAndCreatorEmail(topicId, userEmail)) {
+            topicRepository.topicIsSolved(topicId);
+            commentRepository.commentResolved(commentId);
+            return this.getTopicById(topicId);
+        } else {
+            throw new IllegalArgumentException("This topic has not been created by this user!");
+        }
     }
 
 }
